NON-OFFICIAL/EXPERIMENTAL: add support for .tsx cassette images
This is a follow-up on revision bda300054379 (on the master branch).

I don't (yet?) want to add TSX support in the 'official' openMSX tree,
but I also don't want to hide the work I did. So I'll push it on this
branch.

--

TSX support in openMSX has been somewhat controversial. I won't repeat
the whole discussion. This is only a _very_ brief summary. This is my
personal opinion, not necessarily the opinion of other openMSX team
members.

My main objections against adding TSX support were:
a) TSX is a very COMPLEX file format. And none of the existing
   implementations (openMSX forks) do fully implement it.
b) The openMSX forks with TSX support had several bugs, including buffer
   overflows. These are only bugs and can be solved. But it does reinforce
   the first point.
c) TSX is often seen/proposed as a solution for 'preservation', but it's
   ill suited for that. The users 'bore' and 'Grauw' posted a concise
   reasoning why here (posts of 17-09-2021):
     https://www.msx.org/forum/msx-talk/openmsx/tsx-support-in-openmsx?page=1

This patch does not fully resolve these objections:
* Point a) is (only) partly addressed by only implementing a subset of
  the full TSX specification. (But this is also what the openMSX-forks
  with TSX support already did).
* For point b) I did a detailed review of the code. All buffer overflows
  should be solved (they are now impossible-by-design). And, to the best
  of my knowledge, after some more bug-fixes, the implementation should
  now be compliant with the TSX/TZX specification (at least the subset
  that is actually implemented).
* Point c) remains.
By keeping TSX support out of openMSX I was hoping a 'better' file format
would emerge.

I do understand the value that TSX brings: it adds a semantic
interpretation of the waveform. For certain use cases this can be
useful. But purely from an emulation point of view, we're only
interested in the waveform, not in the interpretation of that waveform.