[Windows] disable IME on main window (Japanese keyboard fix)
add setting "kbd_disable_IME" (default: true)

(For Windows only)
Disable IME in main window.
(Subwindows are not suppressed.)

Basically, if the IME is turned on, there will be problems with key input, so the default is kbd_disable_IME=false.

In case you want to use IME in menus etc. in the future, we have added an option to enable IME.